Remove DCHECK from LegacyFragmentToPaint().
There's code that sets up PaintInfo objects on its own in
SVGObjectPainter, LayoutSVGResourceClipper and CanvasFormattedText,
without specifying a FragmentData ID. None of these should have to deal
with block fragmentation (SVGs and canvases are monolithic, i.e.
unbreakable), but with legacy block fragmentation, we may still create
multiple FragmentData entries for the same object here, if it is too
tall to fit in one whole fragmentainer.

An alternative fix would be to set the fragment ID to the first
FragmentData entry of the object at all these sites, but it doesn't make
much of a difference, and this won't be needed once the legacy engine is
gone, anyway.

Bug: 1238412
Change-Id: I22a8794105c6e106af5b267f34443d8c6ef80e8d